WebIDL/tests/submissions/heycam/Global-002.html

<!DOCTYPE html>
<title>[Global] property lookup</title>
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="utils.js"></script>
<div id="log"></div>

<div id="hidden" style="visibility: hidden; position: absolute; width: 0; height: 0; overflow: hidden">
  <form id="f1" name="f1"></form>
  <form id="f2" name="f2"></form>
</div>

<script>
// 4.4.4.1 Named properties object [[GetOwnProperty]] method
//
// Named properties are exposed as data properties through [[GetOwnProperty]].
// [[GetOwnProperty]] also exposes other (expando) properties on the named
// properties object.

// feature: Window (interface) (with [Global])

var hidden = document.getElementById("hidden");
var form1 = document.getElementById("f1");
var form2 = document.getElementById("f2");

test(function() {
  assert_equals(window.f1, form1);
}, "properties from the named properties object can be looked up on window");

test(function() {
  assert_equals(Window.prototype.f1, form1);
}, "properties from the named properties object can be looked up on Window.prototype");

test(function() {
  var namedPropertiesObject = Object.getPrototypeOf(Window.prototype);
  assert_equals(namedPropertiesObject.f1, form1);
}, "named properties can be looked up on the named properties object");

test(function() {
  var object = window;
  var desc;
  while (object && !desc) {
    desc = Object.getOwnPropertyDescriptor(object, "f1");
    object = Object.getPrototypeOf(object);
  }
  assert_true(!!desc, "named property exposed through [[GetOwnProperty]]");
  assert_descriptor(desc, { writable: false, enumerable: true, configurable: true, value: form1 }, "f1");
}, "named properties are exposed correctly through [[GetOwnProperty]]");

test(function() {
  var namedPropertiesObject = Object.getPrototypeOf(Window.prototype);
  namedPropertiesObject.expando1 = 1;
  assert_equals(window.expando1, 1);
}, "expando properties on the named properties object are exposed through window");

test(function() {
  var namedPropertiesObject = Object.getPrototypeOf(Window.prototype);

  // Ensure there's already a named property "f2".
  assert_property(namedPropertiesObject, "f2", { writable: false, enumerable: true, configurable: true, value: form2 }, "#1 namedPropertiesObject.f2");

  // Define a property with the same name.
  Object.defineProperty(namedPropertiesObject, "f2", { writable: false, enumerable: false, configurable: true, value: "hello" });

  // The expando should now be exposed, since the named property visibility
  // algorithm should prevent the named property from shadowing it.
  assert_property(namedPropertiesObject, "f2", { writable: false, enumerable: false, configurable: true, value: "hello" }, "#2 namedPropertiesObject.f2");

  // Remove the named property.
  form2.parentNode.removeChild(form2);

  // The expando should still be exposed.
  assert_property(namedPropertiesObject, "f2", { writable: false, enumerable: false, configurable: true, value: "hello" }, "#3 namedPropertiesObject.f2");
}, "expando properties with the same as named properties can be set and will shadow named properties");

test(function() {
  var namedPropertiesObject = Object.getPrototypeOf(Window.prototype);

  // Define an expando property.
  Object.defineProperty(namedPropertiesObject, "f3", { writable: false, enumerable: false, configurable: true, value: "hello" });

  // Check that it's there.
  assert_property(namedPropertiesObject, "f3", { writable: false, enumerable: false, configurable: true, value: "hello" }, "#1 namedPropertiesObject.f3");

  // Cause a named property with the same name to start existing.
  var e = document.createElement("form");
  e.name = "f3";
  hidden.appendChild(e);

  // The expando should shadow the named property.
  assert_property(namedPropertiesObject, "f3", { writable: false, enumerable: false, configurable: true, value: "hello" }, "#2 namedPropertiesObject.f3");

  // Remove the named property.
  hidden.removeChild(e);

  // The expando should still be visible.
  assert_property(namedPropertiesObject, "f3", { writable: false, enumerable: false, configurable: true, value: "hello" }, "#3 namedPropertiesObject.f3");
}, "named properties will not shadow an existing expando property");

test(function() {
  var namedPropertiesObject = Object.getPrototypeOf(Window.prototype);

  // Define an expando property.
  Object.defineProperty(namedPropertiesObject, "f4", { writable: false, enumerable: false, configurable: true, value: "hello" });

  // Check that it's there.
  assert_property(namedPropertiesObject, "f4", { writable: false, enumerable: false, configurable: true, value: "hello" }, "#1 namedPropertiesObject.f4");

  // Cause a named property with the same name to start existing.
  var e = document.createElement("form");
  e.name = "f4";
  hidden.appendChild(e);

  // The expando should shadow the named property.
  assert_property(namedPropertiesObject, "f4", { writable: false, enumerable: false, configurable: true, value: "hello" }, "#2 namedPropertiesObject.f4");

  // Remove the expando.
  assert_true(delete namedPropertiesObject.f4, "delete namedPropertiesObject.f4");

  // The named property should now be visible.
  assert_property(namedPropertiesObject, "f4", { writable: false, enumerable: true, configurable: true, value: e }, "#3 namedPropertiesObject.f4");

  // Remove the named property.
  hidden.removeChild(e);

  // There shouldn't be anything there now.
  assert_true(!Object.getOwnPropertyDescriptor(namedPropertiesObject, "f4"), "existence of namedPropertiesObject.f4");
}, "removing an expando property will reveal a shadowed named property");

// XXX We probably shouldn't allow non-configurable properties to be defined
// on the named properties object: https://www.w3.org/Bugs/Public/show_bug.cgi?id=21929
// Once that bug is resolved, we should test this.
</script>
